+++ src/index.html
@@ -9,28 +9,16 @@
 
 <body>
 
-  <header>
-    <!--imagem com bandeiras-->
-    <!--imagem do céu de fundo-->
-  </header>
-
   <main>
     <div id="logoContainer">
       <h2 id="logo"><b>CORREIO ELEGANTE VIRTUAL</b></h2>
       <ul id="menu">
-        <li><a href="message.html" class="menuItem">CIFRAR</a></li>
-        <li><a href="message.html" class="menuItem">DECIFRAR</a></li>
+        <li class="menuItem"><a href="encode.html" class="link">CODIFICAR</a></li>
+        <li class="menuItem"><a href="decode.html" class="link">DECODIFICAR</a></li>
       </ul>
     </div>
   </main>
-    <!-- <nav> 
-    </nav> -->
-  <!-- <footer>
-    <img src="./image/cerca.avif" alt="cerca" class="cerca">
-  </footer> -->
-  <!-- <script src="cipher.js" tyle="module"></script> -->
-  <script src="index.js" type="module"></script>
-  
+
 </body>
 
 </html>
