--- src/message.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="utf-8">
-  <title>Caesar Cipher</title>
-  <link rel="stylesheet" href="style.css" />
-</head>
-
-<body>
-  <header>
-    <!--imagem com bandeiras-->
-  </header>
-
-  <main>
-    <div id="containerCifrar">
-      <!--caixa de mensagem-->
-      <div class="messageContainer">
-        <!--não esquecer de inserir um valor para o atributo ACTION da tag form
-            O Backand tem que estar preparado para receber um POST method-->
-        <form action="">
-          <label for="msg">Mensagem:</label><br>
-          <input type="text" id="conteudo" class="msg">
-          <br>
-          <label for="msg">Mensagem cifrada:</label><br>
-          <input type="text" id="cifrar" class="msg">
-          <br>
-          <input type="number" min="0" id="number">
-          <input type="reset" value="Limpar" class="button">
-          <input type="submit" id="submit" value="Enviar" class="button">
-        </form>
-      </div>
-    </div>
-
-    <div id="containerDecifrar">
-      <div class="messageContainer">
-        <form action="">
-          <label for="msg">Mensagem:</label><br> <!--o nome for tem que ser igual ao do id-->
-          <input type="text" id="message" class="msg">
-          <br>
-          <label for="msg">Mensagem decifrada:</label><br>
-          <input type="text" id="decifrar" class="msg">
-          <br>
-          <input type="number" min="0" id="desloc">
-          <input type="reset" value="Limpar" class="button">
-          <input type="submit" id="send" value="Enviar" class="button">
-        </form>
-      </div>
-    </div>
-  </main>
-
-  <footer>
-    <!--imagem de carta-->
-    <!-- <img src="./image/cerca.avif" alt="cerca" class="cerca"> -->
-  </footer>
-
-  <!-- <script src="cipher.js" tyle="module"></script> -->
-  <script src="index.js" type="module"></script>
-  
-</body>
-</html>
