+++ src/style.css
@@ -1,126 +1,145 @@
 * {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
 }
 
 body {
-    height: 100vh;
-    font-family: cursive;
-    margin: 0;
-    padding: 0;
-    background-color: rgb(2, 2, 83);
-    background-image: url("./image/bandeiras.jpg");
-    background-position: center center;
-    background-repeat: no-repeat;
-    background-size: cover;
+  height: 100vh;
+  font-family: cursive;
+  margin: 0;
+  padding: 0;
+  background-color: white;
+  background-image: url("./image/festaJunina.jpg");
+  background-position: center center;
+  background-repeat: no-repeat;
+  background-size: cover;
 }
 
 main {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: 200px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 200px;
 }
 
 #logoContainer {
-    width: 33%;
-    height: 300px;
-    background-color: rgb(227, 99, 14);
-    box-shadow: .5px 10px 15px rgba(0, 0, 0, 0.529);
-    border-radius: 20px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
+  width: 33%;
+  height: 300px;
+  background-color: rgb(118, 97, 67);
+  box-shadow: .5px 10px 15px rgba(0, 0, 0, 0.645);
+  border-radius: 20px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
 }
 
 #logo {
-    display: flex;
-    position: relative;
-    color: rgb(242, 242, 208);
-    text-align: center;
-    justify-content: center;
-    margin-bottom: 40px;
+  display: flex;
+  position: relative;
+  color: white;
+  text-align: center;
+  justify-content: center;
+  margin-bottom: 50px;
+  font-size: 30px;
+  -webkit-text-stroke-width: 3px;
+  -webkit-text-stroke-color: black;
 }
 
 .menuItem {
-    display: flex;
-    position: relative;
-    color: rgb(242, 242, 208);
-    padding: 15px;
-    margin-left: 25px;
-    text-decoration: none;
-    border-radius: 5px;
-    list-style: none;
-    justify-content: center;
-    float: left;
-    background-color: rgba(146, 63, 8, 0.8);
+  display: flex;
+  position: relative;
+  color: white;
+  padding: 15px;
+  margin-left: 25px;
+  border-radius: 5px;
+  list-style: none;
+  justify-content: center;
+  float: left;
+  background-color: rgba(31, 27, 25, 0.8);
+  font-weight: bolder;
+}
+
+.link {
+  text-decoration: none;
+  color: white;
+  font-size: 14px;
 }
 
 
 #containerCifrar {
-    width: 33%;
-    height: 250px;
-    background-color: rgb(227, 99, 14);
-    border-radius: 20px;
-    box-shadow: .5px 10px 15px rgba(0, 0, 0, 0.529);
-    margin: 25px auto;
-    padding: 25px;
+  width: 33%;
+  height: 300px;
+  background-color: rgb(118, 97, 67);
+  border-radius: 20px;
+  box-shadow: .5px 10px 15px rgba(0, 0, 0, 0.645);
+  margin: 25px auto;
+  padding: 25px;
 }
 
 #containerDecifrar {
-    width: 33%;
-    height: 250px;
-    background-color: rgb(227, 99, 14);
-    border-radius: 20px;
-    box-shadow: .5px 10px 15px rgba(0, 0, 0, 0.396);
-    margin: 25px auto;
-    padding: 25px;
+  width: 33%;
+  height: 300px;
+  background-color: rgb(118, 97, 67);
+  border-radius: 20px;
+  box-shadow: .5px 10px 15px rgba(0, 0, 0, 0.645);
+  margin: 25px auto;
+  padding: 25px;
 
 
 }
 
 .messageContainer {
-    display: flex;
-    position: relative;
-    justify-content: center;
-    align-items: end;
+  display: flex;
+  position: relative;
+  justify-content: center;
+  align-items: end;
 }
 
 label {
-    color: rgb(242, 242, 208);
-    /* font-family: "times new roman"; */
+  color: white;
 }
 
 .msg {
-    width: 300px;
-    height: 50px;
-    border-radius: 5px;
-    border-color: rgb(227, 99, 14);
+  width: 300px;
+  height: 50px;
+  border-radius: 5px;
+  border-color: black;
+  margin-bottom: 25px;
 }
 
-#number{
-    width: 60px;
-    color: black;
-    background-color: white;
-    border-radius: 5px;
-    border-color: white;
+#number {
+  width: 60px;
+  color: black;
+  background-color: white;
+  border-radius: 5px;
+  border-color: white;
 
 }
 
-#desloc{
-    width: 60px;
-    color: black;
-    background-color: white;
-    border-radius: 5px;
-    border-color: white;
+#desloc {
+  width: 60px;
+  color: black;
+  background-color: white;
+  border-radius: 5px;
+  border-color: white;
 }
 
-.button{
-    width: 60px;
-    color: black;
-    background-color: white;
-    border-radius: 5px;
-    border-color: white;
+.button {
+  width: 60px;
+  color: black;
+  background-color: white;
+  border-radius: 5px;
+  border-color: white;
+  font-family: cursive;
+}
+
+.backToMenu {
+  border-radius: 5px;
+  border-color: rgba(31, 27, 25, 0.8);
+  background-color: rgba(31, 27, 25, 0.8);
+  color: white;
+  padding: 10px;
+  font-family: cursive;
 }
